+++ src/Undefinable/index.ts
@@ -7,7 +7,7 @@ export { mapForUndefinable as map } from './map';
 export { mapOrForUndefinable as mapOr } from './mapOr';
 export { mapOrElseForUndefinable as mapOrElse } from './mapOrElse';
 //export { orUndefinable as or} from './or';
-//export { orElseForUndefinable as orElse} from './orElse';
+export { orElseForUndefinable as orElse} from './orElse';
 export { unwrapUndefinable as unwrap } from './unwrap';
 export { unwrapOrFromUndefinable as unwrapOr } from './unwrapOr';
 export { unwrapOrElseFromUndefinable as unwrapOrElse } from './unwrapOrElse';