+++ CHANGELOG.md
@@ -9,6 +9,10 @@
   - We drop to support for `flatMap()` for `Nullable` or `Undefinable`.
     - It's too hard to undarstand that "flatMap" operation for `T | undefined` or `T | null`.
 
+### Enhancement
+
+- You can use `orElse()` function from `option-t/lib/Nullable` or `option-t/lib/Undefinable`. ([#170](https://github.com/saneyuki/option-t.js/pull/170))
+
 
 ## 4.1.0